Add error handler to reduce silent failures.
This shows a warning in the top-right corner of the screen, next to the
version, if any sort of error occurs on the page.

This can be a game error or a script error. It's not possible to
reliably tell where it happens.

Known errors in 1.3.13b are ignored, but the error message format
is not perfectly stable across browsers, so that's best-effort.