Tighten Forager default-job logic
If no workers are assigned, it's possible to end up setting Unemployed as default job,
which the game allows for some reason even though the job isn't visible.
That sinks all normal population and makes the script not consider those workers.

This only seems to actually break if scavengers aren't available on the current planet.

This moves it later in the priority list but removes the worker requirement, and adds it to the emergency fallback.
Not perfect but Forager is the only basic job Racconar have so it makes no real difference other than fixing the bug if no.