Remove currentQuantity check from requestQuantity()
This check shouldn't be there. It was present in about half of the checks before
commit 867e68d0a9aa75d18e4183570232feb5c2cb0e98 replaced this pattern across the codebase.
In retrospect it doesn't make any sense for it to be there at all.

This hopefully fixes a problem where triggers/automech/etc will save up resources for something,
but then the script spends it all on something different than what was saved up for after
the resources are available.